PDf exported for large amount of processed docs
git-svn-id: https://xenitsupport.jira.com/svn/XENMTOA/trunk/move2alf-war@826 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ src/main/java/eu/xenit/move2alf/web/controller/JobController.java
@@ -817,7 +817,7 @@ public class JobController {
 			if(totalTimeInSeconds==0){
 				docsPerSecond = ""+documentListSize;
 			}else{
-				docsPerSecond = ""+documentListSize/totalTimeInSeconds;
+				docsPerSecond = ""+(double)documentListSize/totalTimeInSeconds;
 			}
 		}
 		PagedListHolder pagedListHolder = new PagedListHolder(processedDocuments);
@@ -866,7 +866,7 @@ public class JobController {
 			if(totalTimeInSeconds==0){
 				docsPerSecond = ""+documentListSize;
 			}else{
-				docsPerSecond = ""+documentListSize/totalTimeInSeconds;
+				docsPerSecond = ""+(double)documentListSize/totalTimeInSeconds;
 			}
 		}
 		PagedListHolder pagedListHolder = new PagedListHolder(processedDocuments);
@@ -916,7 +916,7 @@ public class JobController {
 			if(totalTimeInSeconds==0){
 				docsPerSecond = ""+documentListSize;
 			}else{
-				docsPerSecond = ""+documentListSize/totalTimeInSeconds;
+				docsPerSecond = ""+(double)documentListSize/totalTimeInSeconds;
 			}
 		}
 		
@@ -961,7 +961,7 @@ public class JobController {
 			if(totalTimeInSeconds==0){
 				docsPerSecond = ""+documentListSize;
 			}else{
-				docsPerSecond = ""+documentListSize/totalTimeInSeconds;
+				docsPerSecond = ""+(double)documentListSize/totalTimeInSeconds;
 			}
 		}
